Scope sonatypeCredentialHost to the build instead of projects.
Previously, the `sonatypeCredentialHost` setting had to be redefined
inside every project even if the sbt-sonatype readme documents that
the setting can be configured once via `ThisBuild /` at the root of
build.sbt. This commit moves the definition of that setting to the build
scope so that the behavior is consistent with the documentation.

Related https://github.com/olafurpg/sbt-ci-release/issues/173