Give helpful advice when profile missing on host
This updates the error message given when a profile is not found on a host (note Sonatype currently has 2
OSS hosts - `oss.sonatype.org` & `s01.oss.sonatype.org` and any profile will only exist on _one_ of the
hosts).

There can be two possible causes when a profile is not found on a host:

1. The profile name is misspelt/incorrect (eg 'com.gnu' or 'john.smith' instead of 'com.gu')
2. The profile is actually on the _other_ Sonatype OSS host

The original error message really only indicated the possibility of the _1st_ problem, but the 2nd issue
is increasingly likely - since February 2021 new profiles have _only_ been registered on the new host
`s01.oss.sonatype.org` (https://central.sonatype.org/news/20210223_new-users-on-s01/), and `sbt-sonatype`
defaults to using the other host, `oss.sonatype.org`.

The new error message aims to make the user aware of the 2 possible causes of the 'missing profile'
problem, and offer advice on how the 2nd problem could be corrected.

Before:

> Profile com.gu is not found. Check your sonatypeProfileName setting in build.sbt

After:

> Profile com.gu is not found on oss.sonatype.org. In your sbt settings, check your sonatypeProfileName and sonatypeCredentialHost (try s01.oss.sonatype.org?)"